backend/ml_requirements.txt
# ML System Requirements for PantryPal
# Install with: pip install -r ml_requirements.txt

# Core ML Libraries
lightgbm>=4.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Deep Learning & Embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Vector Search
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support

# Time Series Forecasting
prophet>=1.1.0  # Requires pystan
statsmodels>=0.14.0

# Model Explainability
shap>=0.43.0

# Text Processing
rapidfuzz>=3.0.0

# Optimization
ortools>=9.7.0

# Model Registry & Tracking
mlflow>=2.8.0
bentoml>=1.1.0  # Alternative to MLflow

# Data Validation & Quality
great-expectations>=0.17.0  # Optional: data quality checks
pandas-profiling>=4.0.0  # Optional: data exploration

# Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
joblib>=1.3.0  # Model serialization
tqdm>=4.66.0  # Progress bars
python-dotenv>=1.0.0  # Environment variables

# API Integration (if serving models via API)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Database (already in main requirements.txt, but listed for reference)
# sqlalchemy>=2.0.0
# asyncpg>=0.29.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Optional: GPU Acceleration
# cupy>=12.0.0  # GPU-accelerated numpy
# cudf>=23.0.0  # GPU-accelerated pandas (RAPIDS)

# Optional: Advanced Time Series
# neuralprophet>=0.6.0  # Neural network-based Prophet
# sktime>=0.24.0  # Advanced time series ML

# Optional: AutoML
# optuna>=3.4.0  # Hyperparameter tuning
# auto-sklearn>=0.15.0  # AutoML framework

# Optional: Distributed Computing
# dask[complete]>=2023.10.0  # Parallel computing
# ray[tune]>=2.8.0  # Distributed hyperparameter tuning
